test: ☑️ refactor post model tests
