+++ tests/api/main.rs
@@ -1,4 +1,5 @@
 #![warn(clippy::all, clippy::pedantic)]
 
 mod helpers;
+mod model;
 mod startup;
