+++ tests/api/model/mod.rs
@@ -1,0 +1,1 @@
+mod post;
